+++ src/Aws/ElastiCache/Exception/ReservedCacheNodeAlreadyExistsException.php
@@ -17,6 +17,6 @@
 namespace Aws\ElastiCache\Exception;
 
 /**
- * This user already has a reservation with the given identifier.
+ * You already have a reservation with the given identifier.
  */
 class ReservedCacheNodeAlreadyExistsException extends ElastiCacheException {}